pyproject.toml
[tool.poetry]
name = "py_name_entity_recognition"
version = "0.4.0"
description = "LLM-based Named Entity Recognition using LangChain and LangGraph"
authors = ["Gowtham Rao <rao@ohdsi.org>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/GowthamRao/py_name_entity_recognition"
repository = "https://github.com/GowthamRao/py_name_entity_recognition"
keywords = ["ner", "named entity recognition", "llm", "langchain", "langgraph", "nlp"]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.0"
langchain-community = "^0.2.0"
langchain-openai = "^0.1.0"
langchain-anthropic = "^0.1.0"
langgraph = "^0.1.0"
pydantic = "^2.5.0"
spacy = "^3.7.0"
loguru = "^0.7.0"
tenacity = "^8.2.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
pandas = "^2.0"
datasets = "^2.14.0"
Jinja2 = "^3.1.0"
# For model providers
openai = "^1.3.0"
anthropic = "^0.23.0"
google-generativeai = "^0.3.0"
ollama = "^0.1.0"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.6.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.1"
black = "^23.11.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]
